refactor(fathers page): extract useMouseTrail in a separate hook file in Background
